Refactor verse segmentation
